Minor: renaming, comments and move
