+++ tianshou/data/types.py
@@ -54,3 +54,13 @@ class FQFBatchProtocol(ModelOutputBatchProtocol):
 
     fractions: torch.Tensor
     quantiles_tau: torch.Tensor
+
+
+class BatchWithAdvantagesProtocol(BatchWithReturnsProtocol):
+    """Contains estimated advantages and values.
+
+    Returns are usually computed from GAE of advantages by adding the value.
+    """
+
+    adv: torch.Tensor
+    v_s: torch.Tensor
